chore: separated azure unit tests into separate files. fixed some of its tests.
